SQL Connection

+++ Flockbuster.Domain/Models/Movies.cs
@@ -11,7 +11,7 @@ namespace Flockbuster.Domain.Models
         public int Id { get; set; }
         public string Title { get; set; } = string.Empty;
         public int RequiredAge { get; set; }
-        public long TTW { get; set; }
+        public int TTW { get; set; }
         public string RelaseDate { get; set; } = string.Empty;
         public int Price { get; set; }
         public List<Genre> Genres { get; set; }
